feat: Moved queue-related code to queue/index.ts
